+++ CHANGELOG.md
@@ -5,6 +5,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+
+## [1.1.1] - 2018-06-11
 ### Added
 - invocation of "markOrderAsShopgate" pipeline from frontend code by subscribing to the checkoutSuccess event